+++ R/tar_prune.R
@@ -1,18 +1,20 @@
 #' @title Remove targets that are no longer part of the pipeline.
 #' @export
 #' @family clean
+#' @seealso tar_prune_inspect
 #' @description Remove target values from `_targets/objects/` and the cloud
 #'   and remove target metadata from `_targets/meta/meta`
 #'   for targets that are no longer part of the pipeline.
 #' @details
-#'   This is useful if you recently worked through
+#'   `tar_prune()` is useful if you recently worked through
 #'   multiple changes to your project and are now trying to
 #'   discard irrelevant data while keeping the results that still matter.
 #'   Global objects and local files with `format = "file"` outside the
 #'   data store are unaffected. Also removes `_targets/scratch/`,
 #'   which is only needed while [tar_make()], [tar_make_clustermq()],
-#'   or [tar_make_future()] is running.
-#' @return `NULL` except if `callr_function = callr::r_bg()`, in which case
+#'   or [tar_make_future()] is running. To list the targets that will be
+#'   pruned without actually removing anything, use [tar_prune_list()].
+#' @return `NULL` except if `callr_function` is `callr::r_bg`, in which case
 #'   a handle to the `callr` background process is returned. Either way,
 #'   the value is invisibly returned.
 #' @inheritParams tar_validate